backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google AI (Gemini Vision)
google-generativeai==0.3.2

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Blockchain
web3>=7.0.0
eth-account>=0.11.0

# Image processing
Pillow>=10.4.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Google Cloud Storage
google-cloud-storage==2.14.0

# SpoonOS SDK
spoon-ai-sdk>=0.3.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
